requirements.txt
pandas==1.5.3
openpyxl==3.0.10
scipy==1.10.1
scikit-learn==1.2.1
numpy==1.24.2
torch==1.13.1
torchvision==0.14.1
Pillow==9.4.0